feat: add --list-languages option
It prints installed language files, which are under
TTYPER_CONFIG_DIR/language, separated by newline and then exits.

Currently we need to find the directory to know which languages are
installed. But it is bothersome because we have to read README or source
code to know where the directory is on. So add --list-languages option
to list installed languages by this command, which has known where the
directory is on.